+++ crates/wdl-analysis/src/handlers.rs
@@ -10,6 +10,7 @@ use crate::types::v1::EvaluationContext;
 
 mod common;
 mod completions;
+mod document_symbol;
 mod find_all_references;
 mod goto_definition;
 mod hover;
@@ -17,6 +18,7 @@ mod rename;
 mod semantic_tokens;
 
 pub use completions::*;
+pub use document_symbol::*;
 pub use find_all_references::*;
 pub use goto_definition::*;
 pub use hover::*;