+++ crates/wdl-analysis/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+#### Added
+
+* Added document symbols support for the WDL Language Server ([#582](https://github.com/stjude-rust-labs/wdl/pull/582)).
+
 ## 0.12.0 - 08-13-2025
 
 * Added semantic highlighting support for the WDL Language Server ([#569](https://github.com/stjude-rust-labs/wdl/pull/569)).