+++ crates/wdl-lsp/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+#### Added
+
+* Added support for document symbols ([#582](https://github.com/stjude-rust-labs/wdl/pull/582)).
+
 ## 0.12.0 - 08-13-2025
 
 #### Added